requirements.txt
torch
torchvision
numpy
scikit-learn
joblib
matplotlib
Pillow
Gradio
